Factor out an it helper
